blessed-rewrite branch initial commit
- New and updated logo
- Renamed index.py to old_index.py and game.py to old_game.py
- Updated requirements.txt
- First part of the rewrite: brand new index.py
- Updated README.md
- Languages! Hah!
