+++ assets/logo.txt
@@ -1,12 +1,12 @@
           __                __ __          __                  __     __                    
          |  \              |  \  \        |  \                |  \   |  \                   
-  _______| @@____   ______ | @@ @@ ______ | @@____  __    __ _| @@_  | @@____  ______ ____  
- /       \ @@    \ /      \| @@ @@/      \| @@    \|  \  |  \   @@ \ | @@    \|      \    \ 
-|  @@@@@@@ @@@@@@@\  @@@@@@\ @@ @@  @@@@@@\ @@@@@@@\ @@  | @@\@@@@@@ | @@@@@@@\ @@@@@@\@@@@\
- \@@    \| @@  | @@ @@    @@ @@ @@ @@   \@@ @@  | @@ @@  | @@ | @@ __| @@  | @@ @@ | @@ | @@
- _\@@@@@@\ @@  | @@ @@@@@@@@ @@ @@ @@     | @@  | @@ @@__/ @@ | @@|  \ @@  | @@ @@ | @@ | @@
-|       @@ @@  | @@\@@     \ @@ @@ @@     | @@  | @@\@@    @@  \@@  @@ @@  | @@ @@ | @@ | @@
- \@@@@@@@ \@@   \@@ \@@@@@@@\@@\@@\@@      \@@   \@@_\@@@@@@@   \@@@@ \@@   \@@\@@  \@@  \@@
+  _______| ##____   ______ | @@ @@ ______ | ##____  __    __ _| @@_  | ##____  ______ ____  
+ /       \ ##    \ /      \| @@ @@/      \| ##    \|  \  |  \   @@ \ | ##    \|      \    \ 
+|  @@@@@@@ #######\  $$$$$$\ @@ @@  @@@@@@\ #######\ @@  | @@\@@@@@@ | #######\ @@@@@@\@@@@\
+ \@@    \| ##  | ## $$    $$ @@ @@ @@   \@@ ##  | ## @@  | @@ | @@ __| ##  | ## @@ | @@ | @@
+ _\@@@@@@\ ##  | ## $$$$$$$$ @@ @@ @@     | ##  | ## @@__/ @@ | @@|  \ ##  | ## @@ | @@ | @@
+|       @@ ##  | ##\$$     \ @@ @@ @@     | ##  | ##\@@    @@  \@@  @@ ##  | ## @@ | @@ | @@
+ \@@@@@@@ \##   \## \$$$$$$$\@@\@@\@@      \##   \##_\@@@@@@@   \@@@@ \##   \##\@@  \@@  \@@
                                                    |  \__| @@                               
                                                     \@@    @@                               
                                                      \@@@@@@
